requirements.txt
django==5.2
Pillow==11.3.0
django-admin-sortable2==2.2.8
